+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased] - TBD
+- Added `--direction` flag to define a diagram direction: "top to bottom" or "left to right".
 - Fixed the current working directory issue.
 
 ## [v1.3.0] (2020-11-12)